+++ ton_client/src/debot/tests.rs
@@ -91,6 +91,17 @@ const TERMINAL_ABI: &str = r#"
 			"outputs": [
 				{"name":"value","type":"int256"}
 			]
+		},
+        {
+			"name": "input",
+			"inputs": [
+				{"name":"answerId","type":"uint32"},
+				{"name":"prompt","type":"bytes"},
+				{"name":"multiline","type":"bool"}
+			],
+			"outputs": [
+				{"name":"value","type":"bytes"}
+			]
 		}
 	],
 	"data": [],
@@ -151,7 +162,16 @@ impl Terminal {
                 let _ = self.print(answer_id, prompt);
                 // use test return value here.
                 (answer_id, json!({"value": 1}))
-            }
+            },
+            "input" => {
+                let answer_id = decode_abi_number::<u32>(args["answerId"].as_str().unwrap()).unwrap();
+                let prompt = hex::decode(args["prompt"].as_str().unwrap()).unwrap();
+                let message = std::str::from_utf8(&prompt).unwrap();
+                let _ = args["multiline"].as_bool().unwrap();
+                self.print(answer_id, message);
+                let value = "testinput";
+                (answer_id, json!({ "value": hex::encode(value.as_bytes()) }) )
+            },
             _ => panic!("interface function not found"),
         }
     }
@@ -187,10 +207,15 @@ struct BrowserData {
     pub terminal: Mutex<Terminal>,
     pub echo: Echo,
     pub bots: Mutex<HashMap<String, RegisteredDebot>>,
+    pub info: DebotInfo,
 }
 
 impl TestBrowser {
-    async fn fetch_debot(client: Arc<TestClient>, state: Arc<BrowserData>, address: String, start_function: &str) -> RegisteredDebot {
+    async fn fetch_debot(
+        client: Arc<TestClient>,
+        state: Arc<BrowserData>,
+        address: String,
+    ) -> RegisteredDebot {
         let state_copy = state.clone();
         let client_copy = client.clone();
         let callback = move |params, response_type| {
@@ -219,21 +244,38 @@ impl TestBrowser {
         };
 
         let handle: RegisteredDebot = client.request_async_callback(
-            start_function,
-            ParamsOfStart { address: address.clone() },
+            "debot.init",
+            ParamsOfInit { address: address.clone() },
             callback
         ).await.unwrap();
 
         let handle_copy = RegisteredDebot {
             debot_handle: handle.debot_handle.clone(),
             debot_abi: handle.debot_abi.clone(),
+            info: handle.info.clone(),
         };
         state.bots.lock().await.insert(address.clone(), handle_copy);
         handle
     }
 
-    pub async fn execute_from_state(client: Arc<TestClient>, state: Arc<BrowserData>, start_function: &str) {
-        let handle = Self::fetch_debot(client.clone(), state.clone(), state.address.clone(), start_function).await;
+    pub async fn execute_from_state(client: Arc<TestClient>, state: Arc<BrowserData>, call_start: bool) {
+        if call_start {
+            let res: ResultOfFetch = client.request_async(
+                "debot.fetch",
+                ParamsOfFetch { address: state.address.clone() },
+            ).await.unwrap();
+            assert_eq!(res.info, state.info);
+        }
+        let handle = Self::fetch_debot(client.clone(), state.clone(), state.address.clone()).await;
+
+        if call_start {
+            let _: () = client.request_async(
+                "debot.start",
+                ParamsOfStart {
+                    debot_handle: handle.debot_handle.clone(),
+                }
+            ).await.unwrap();
+        }
 
         while !state.finished.load(Ordering::Relaxed) {
             Self::handle_message_queue(client.clone(), state.clone()).await;
@@ -284,17 +326,35 @@ impl TestBrowser {
         keys: KeyPair,
         steps: Vec<DebotStep>,
         terminal_outputs: Vec<String>,
+        abi: String,
     ) {
-        Self::execute_with_func(client, address, keys, steps, terminal_outputs, "debot.start").await;
+        let mut info = DebotInfo::default();
+        info.dabi = Some(abi);
+        let state = Arc::new(BrowserData {
+            current: Mutex::new(Default::default()),
+            next: Mutex::new(steps),
+            client: client.clone(),
+            keys,
+            address: address.clone(),
+            finished: AtomicBool::new(false),
+            switch_started: AtomicBool::new(false),
+            msg_queue: Mutex::new(Default::default()),
+            terminal: Mutex::new(Terminal::new(terminal_outputs)),
+            echo: Echo::new(),
+            bots: Mutex::new(HashMap::new()),
+            info,
+        });
+
+        Self::execute_from_state(client, state, true).await
     }
 
-    pub async fn execute_with_func(
+    pub async fn execute_with_info(
         client: Arc<TestClient>,
         address: String,
         keys: KeyPair,
         steps: Vec<DebotStep>,
         terminal_outputs: Vec<String>,
-        entry_function: &str,
+        info: DebotInfo,
     ) {
         let state = Arc::new(BrowserData {
             current: Mutex::new(Default::default()),
@@ -308,9 +368,10 @@ impl TestBrowser {
             terminal: Mutex::new(Terminal::new(terminal_outputs)),
             echo: Echo::new(),
             bots: Mutex::new(HashMap::new()),
+            info,
         });
 
-        Self::execute_from_state(client, state, entry_function).await
+        Self::execute_from_state(client, state, true).await
     }
 
     async fn process_notification(state: &BrowserData, params: ParamsOfAppDebotBrowser) {
@@ -339,9 +400,11 @@ impl TestBrowser {
     }
 
     fn call_execute_boxed(
-        client: Arc<TestClient>, state: Arc<BrowserData>, start_function: &'static str
+        client: Arc<TestClient>,
+        state: Arc<BrowserData>,
+        call_start: bool,
     ) -> BoxFuture<'static, ()> {
-        Self::execute_from_state(client, state, start_function).boxed()
+        Self::execute_from_state(client, state, call_start).boxed()
     }
 
     async fn process_call(client: Arc<TestClient>, state: &BrowserData, params: ParamsOfAppDebotBrowser) -> ResultOfAppDebotBrowser {
@@ -378,8 +441,9 @@ impl TestBrowser {
                     terminal: Mutex::new(Terminal::new(vec![])),
                     echo: Echo::new(),
                     bots: Mutex::new(HashMap::new()),
+                    info: Default::default(),
                 });
-                Self::call_execute_boxed(client, state, "debot.fetch").await;
+                Self::call_execute_boxed(client, state, false).await;
                 ResultOfAppDebotBrowser::InvokeDebot
             },
             ParamsOfAppDebotBrowser::Approve {activity} => {
@@ -447,7 +511,6 @@ impl TestBrowser {
                         client.clone(),
                         state.clone(),
                         dest_addr.to_owned(),
-                        "debot.fetch",
                     ).await;
 
                 }
@@ -492,6 +555,7 @@ struct DebotData {
     debot_addr: String,
     target_addr: String,
     keys: KeyPair,
+    abi: String,
 }
 
 async fn init_debot(client: Arc<TestClient>) -> DebotData {
@@ -572,7 +636,8 @@ async fn init_debot(client: Arc<TestClient>) -> DebotData {
     let data = DebotData {
         debot_addr,
         target_addr,
-        keys
+        keys,
+        abi: debot_abi.json_string().unwrap(),
     };
     *debot = Some(data.clone());
     data
@@ -606,7 +671,7 @@ async fn init_debot2(client: Arc<TestClient>) -> DebotData {
         Signer::Keys { keys: keys.clone() },
     ).await.unwrap();
     let target_addr = String::new();
-    DebotData { debot_addr, target_addr, keys }
+    DebotData { debot_addr, target_addr, keys, abi: debot_abi.json_string().unwrap(), }
 }
 
 async fn init_debot4(client: Arc<TestClient>) -> DebotData {
@@ -662,18 +727,23 @@ async fn init_debot4(client: Arc<TestClient>) -> DebotData {
     DebotData {
         debot_addr,
         target_addr,
-        keys
+        keys,
+        abi: debot_abi.json_string().unwrap(),
     }
 }
 
 async fn init_debot3(client: Arc<TestClient>) -> DebotData {
+    init_simple_debot(client, TEST_DEBOT3).await
+}
+
+async fn init_simple_debot(client: Arc<TestClient>, name: &str) -> DebotData {
     let keys = client.generate_sign_keys();
-    let debot_abi = TestClient::abi(TEST_DEBOT3, Some(2));
+    let debot_abi = TestClient::abi(name, Some(2));
 
     let call_set = CallSet::some_with_function("constructor");
     let deploy_debot_params = ParamsOfEncodeMessage {
         abi: debot_abi.clone(),
-        deploy_set: DeploySet::some_with_tvc(TestClient::tvc(TEST_DEBOT3, Some(2))),
+        deploy_set: DeploySet::some_with_tvc(TestClient::tvc(name, Some(2))),
         signer: Signer::Keys { keys: keys.clone() },
         processing_try_index: None,
         address: None,
@@ -691,10 +761,11 @@ async fn init_debot3(client: Arc<TestClient>) -> DebotData {
         debot_addr,
         target_addr: String::new(),
         keys,
+        abi: debot_abi.json_string().unwrap(),
     }
 }
 
-async fn init_debot5(client: Arc<TestClient>, count: u32) -> String {
+async fn init_debot5(client: Arc<TestClient>, count: u32) -> (String, String) {
     let debot_abi = TestClient::abi(TEST_DEBOT5, Some(2));
     let debot5_tvc = TestClient::tvc(TEST_DEBOT5, Some(2));
 
@@ -735,10 +806,10 @@ async fn init_debot5(client: Arc<TestClient>, count: u32) -> String {
             ).await.unwrap();
         }
     }
-    addrs[0].clone()
+    (addrs[0].clone(), debot_abi.json_string().unwrap())
 }
 
-async fn init_debot_pair(client: Arc<TestClient>, debot1: &str, debot2: &str) -> (String, String) {
+async fn init_debot_pair(client: Arc<TestClient>, debot1: &str, debot2: &str) -> (String, String, String) {
     let keys = client.generate_sign_keys();
     let debot1_abi = TestClient::abi(debot1, Some(2));
     let debot2_abi = TestClient::abi(debot2, Some(2));
@@ -789,8 +860,20 @@ async fn init_debot_pair(client: Arc<TestClient>, debot1: &str, debot2: &str) ->
 
     let (_, _) = futures::join!(future1, future2);
 
-    (debot1_addr, debot2_addr)
+    (debot1_addr, debot2_addr, debot1_abi.json_string().unwrap())
+}
 
+async fn init_hello_debot(client: Arc<TestClient>) -> DebotData {
+    let data = init_simple_debot(client.clone(), "helloDebot").await;
+    let abi = Abi::Contract(serde_json::from_str(&data.abi).unwrap());
+    let _ = client.net_process_function(
+        data.debot_addr.clone(),
+        abi,
+        "setIcon",
+        json!({ "icon": TestClient::icon("helloDebot", Some(2)) }),
+        Signer::Keys { keys: data.keys.clone() },
+    ).await.unwrap();
+    data
 }
 
 const EXIT_CHOICE: u8 = 9;
@@ -798,7 +881,7 @@ const EXIT_CHOICE: u8 = 9;
 #[tokio::test(core_threads = 2)]
 async fn test_debot_goto() {
     let client = std::sync::Arc::new(TestClient::new());
-    let DebotData { debot_addr, target_addr: _, keys } = init_debot(client.clone()).await;
+    let DebotData { debot_addr, target_addr: _, keys, abi } = init_debot(client.clone()).await;
 
     let steps = json!([
         { "choice": 1, "inputs": [], "outputs": ["Test Goto Action"] },
@@ -811,13 +894,14 @@ async fn test_debot_goto() {
         keys.clone(),
         serde_json::from_value(steps).unwrap(),
         vec![],
+        abi
     ).await;
 }
 
 #[tokio::test(core_threads = 2)]
 async fn test_debot_print() {
     let client = std::sync::Arc::new(TestClient::new());
-    let DebotData { debot_addr, target_addr, keys } = init_debot(client.clone()).await;
+    let DebotData { debot_addr, target_addr, keys, abi } = init_debot(client.clone()).await;
 
     let steps = json!([
         { "choice": 2, "inputs": [], "outputs": ["Test Print Action", "test2: instant print", "test instant print"] },
@@ -832,13 +916,14 @@ async fn test_debot_print() {
         keys.clone(),
         serde_json::from_value(steps).unwrap(),
         vec![],
+        abi
     ).await;
 }
 
 #[tokio::test(core_threads = 2)]
 async fn test_debot_runact() {
     let client = std::sync::Arc::new(TestClient::new());
-    let DebotData { debot_addr, target_addr: _, keys } = init_debot(client.clone()).await;
+    let DebotData { debot_addr, target_addr: _, keys, abi } = init_debot(client.clone()).await;
 
     let steps = json!([
         { "choice": 3, "inputs": [], "outputs": ["Test Run Action"] },
@@ -854,14 +939,15 @@ async fn test_debot_runact() {
         debot_addr,
         keys,
         serde_json::from_value(steps).unwrap(),
-        vec![]
+        vec![],
+        abi
     ).await;
 }
 
 #[tokio::test(core_threads = 2)]
 async fn test_debot_run_method() {
     let client = std::sync::Arc::new(TestClient::new());
-    let DebotData { debot_addr, target_addr: _, keys } = init_debot(client.clone()).await;
+    let DebotData { debot_addr, target_addr: _, keys, abi } = init_debot(client.clone()).await;
 
     let steps = json!([
         { "choice": 4, "inputs": [], "outputs": ["Test Run Method Action"] },
@@ -875,14 +961,15 @@ async fn test_debot_run_method() {
         debot_addr,
         keys,
         serde_json::from_value(steps).unwrap(),
-        vec![]
+        vec![],
+        abi
     ).await;
 }
 
 #[tokio::test(core_threads = 2)]
 async fn test_debot_send_msg() {
     let client = std::sync::Arc::new(TestClient::new());
-    let DebotData { debot_addr, target_addr: _, keys } = init_debot(client.clone()).await;
+    let DebotData { debot_addr, target_addr: _, keys, abi } = init_debot(client.clone()).await;
 
     let steps = json!([
         { "choice": 5, "inputs": [], "outputs": ["Test Send Msg Action"] },
@@ -898,13 +985,14 @@ async fn test_debot_send_msg() {
         keys,
         serde_json::from_value(steps).unwrap(),
         vec![],
+        abi
     ).await;
 }
 
 #[tokio::test(core_threads = 2)]
 async fn test_debot_invoke_debot() {
     let client = std::sync::Arc::new(TestClient::new());
-    let DebotData { debot_addr, target_addr: _, keys } = init_debot(client.clone()).await;
+    let DebotData { debot_addr, target_addr: _, keys, abi } = init_debot(client.clone()).await;
 
     let steps = json!([
         { "choice": 6, "inputs": [debot_addr.clone()], "outputs": ["Test Invoke Debot Action", "enter debot address:"] },
@@ -924,14 +1012,15 @@ async fn test_debot_invoke_debot() {
         debot_addr,
         keys,
         serde_json::from_value(steps).unwrap(),
-        vec![]
+        vec![],
+        abi
     ).await;
 }
 
 #[tokio::test(core_threads = 2)]
 async fn test_debot_engine_calls() {
     let client = std::sync::Arc::new(TestClient::new());
-    let DebotData { debot_addr, target_addr: _, keys } = init_debot(client.clone()).await;
+    let DebotData { debot_addr, target_addr: _, keys, abi } = init_debot(client.clone()).await;
 
     let steps = json!([
         { "choice": 7, "inputs": [], "outputs": ["Test Engine Calls"] },
@@ -948,14 +1037,15 @@ async fn test_debot_engine_calls() {
         debot_addr,
         keys,
         serde_json::from_value(steps).unwrap(),
-        vec![]
+        vec![],
+        abi
     ).await;
 }
 
 #[tokio::test(core_threads = 2)]
 async fn test_debot_interface_call() {
     let client = std::sync::Arc::new(TestClient::new());
-    let DebotData { debot_addr, target_addr: _, keys } = init_debot(client.clone()).await;
+    let DebotData { debot_addr, target_addr: _, keys, abi } = init_debot(client.clone()).await;
 
     let steps = json!([
         { "choice": 8, "inputs": [], "outputs": ["", "test1 - call interface"] },
@@ -968,13 +1058,14 @@ async fn test_debot_interface_call() {
         keys,
         serde_json::from_value(steps).unwrap(),
         vec![],
+        abi
     ).await;
 }
 
 #[tokio::test(core_threads = 2)]
 async fn test_debot_inner_interfaces() {
     let client = std::sync::Arc::new(TestClient::new());
-    let DebotData { debot_addr, target_addr: _, keys } = init_debot3(client.clone()).await;
+    let DebotData { debot_addr, target_addr: _, keys, abi } = init_debot3(client.clone()).await;
 
     let steps = serde_json::from_value(json!([])).unwrap();
     TestBrowser::execute(
@@ -998,13 +1089,14 @@ async fn test_debot_inner_interfaces() {
             format!("test hex decode passed"),
             format!("test base64 decode passed"),
         ],
+        abi
     ).await;
 }
 
 #[tokio::test(core_threads = 2)]
 async fn test_debot_4() {
     let client = std::sync::Arc::new(TestClient::new());
-    let DebotData { debot_addr, target_addr, keys } = init_debot4(client.clone()).await;
+    let DebotData { debot_addr, target_addr, keys, abi } = init_debot4(client.clone()).await;
     let target_abi = TestClient::abi(TEST_DEBOT_TARGET, Some(2));
 
     let target_boc = download_account(&client, &target_addr).await.expect("account must exist");
@@ -1031,6 +1123,7 @@ async fn test_debot_4() {
             format!("Transaction succeeded"),
             format!("setData2(129)"),
         ],
+        abi
     ).await;
 
     let target_boc = download_account(&client, &target_addr).await.expect("account must exist");
@@ -1056,7 +1149,7 @@ async fn test_debot_4() {
 #[tokio::test(core_threads = 2)]
 async fn test_debot_msg_interface() {
     let client = std::sync::Arc::new(TestClient::new());
-    let DebotData { debot_addr, target_addr: _, keys } = init_debot2(client.clone()).await;
+    let DebotData { debot_addr, target_addr: _, keys, abi } = init_debot2(client.clone()).await;
     let debot_abi = TestClient::abi(TEST_DEBOT2, Some(2));
     let counter = 10;
     let counter_after = 15;
@@ -1081,6 +1174,7 @@ async fn test_debot_msg_interface() {
             format!("Increment succeeded"),
             format!("counter={}", counter_after),
         ],
+        abi
     ).await;
 
     assert_get_method(
@@ -1096,7 +1190,7 @@ async fn test_debot_msg_interface() {
 #[tokio::test(core_threads = 2)]
 async fn test_debot_invoke_msgs() {
     let client = std::sync::Arc::new(TestClient::new());
-    let (debot1, _) = init_debot_pair(client.clone(), TEST_DEBOTA, TEST_DEBOTB).await;
+    let (debot1, _, abi) = init_debot_pair(client.clone(), TEST_DEBOTA, TEST_DEBOTB).await;
     let keys = client.generate_sign_keys();
 
     let steps = serde_json::from_value(json!([])).unwrap();
@@ -1110,6 +1204,7 @@ async fn test_debot_invoke_msgs() {
             format!("DebotB receives question: What is your name?"),
             format!("DebotA receives answer: My name is DebotB"),
         ],
+        abi
     ).await;
 }
 
@@ -1119,8 +1214,7 @@ async fn test_debot_invoke_msgs() {
 async fn test_debot_sdk_get_accounts_by_hash() {
     let client = std::sync::Arc::new(TestClient::new());
     let count = 6;
-    let debot = init_debot5(client.clone(), count).await;
-
+    let (debot, abi) = init_debot5(client.clone(), count).await;
     let steps = serde_json::from_value(json!([])).unwrap();
     TestBrowser::execute(
         client.clone(),
@@ -1128,6 +1222,39 @@ async fn test_debot_sdk_get_accounts_by_hash() {
         KeyPair::default(),
         steps,
         vec![ format!("{} contracts.", count) ],
+        abi
+    ).await;
+}
+
+#[tokio::test(core_threads = 2)]
+async fn test_debot_getinfo() {
+    let client = std::sync::Arc::new(TestClient::new());
+    let DebotData { debot_addr, target_addr: _, keys, abi } = init_hello_debot(client.clone()).await;
+    let icon = TestClient::icon("helloDebot", Some(2));
+    let steps = serde_json::from_value(json!([])).unwrap();
+    TestBrowser::execute_with_info(
+        client.clone(),
+        debot_addr.clone(),
+        keys,
+        steps,
+        vec![
+            format!("Hello, World!"),
+            format!("How is it going?"),
+            format!("You have entered \"testinput\""),
+        ],
+        DebotInfo {
+            name: Some("HelloWorld".to_owned()),
+            version: Some("0.2.0".to_owned()),
+            publisher: Some("TON Labs".to_owned()),
+            key: Some("Start develop DeBot from here".to_owned()),
+            author: Some("TON Labs".to_owned()),
+            support: Some("0:841288ed3b55d9cdafa806807f02a0ae0c169aa5edfe88a789a6482429756a94".to_owned()),
+            hello: Some("Hello, i am a HelloWorld DeBot.".to_owned()),
+            language: Some("en".to_owned()),
+            dabi: Some(abi),
+            icon: Some(icon),
+            interfaces: vec!["0x8796536366ee21852db56dccb60bc564598b618c865fc50c8b1ab740bba128e3".to_owned()],
+        },
     ).await;
 }